+++ CHANGELOG.md
@@ -21,6 +21,8 @@ See [CHANGELOG - 6to5](CHANGELOG-6to5.md) for the pre-4.0.0 version changelog.
 
     * `babel-types`: Implement validation for `name` field in `Identifier` builder. It's limited to validating that it's a `spec.IdentifierName` since `ast.Identifier` is used to represent those as well. So in cases where the value must be a `spec.Identifier`, this validation will fail to error.
 
+    * `babel-types`: Add `isSpecIdentifier()` as new name for `isValidIdentifier()`.
+
   * **Internal**
 
     * `babel-types`: Start test suite.